Chroma fixes for Pixel=UShort; Bitrate issues sorted out, bit mismatch issues with Pixel=Short

+++ source/Lib/TLibEncoder/TEncSearch.cpp
@@ -1359,7 +1359,7 @@ Void TEncSearch::xIntraCodingChromaBlk(TComDataCU* pcCU,
         {
             for (UInt uiX = 0; uiX < uiWidth; uiX++)
             {
-                pReco[uiX] = ClipC(static_cast<Short>(pPred[uiX]) + pResi[uiX]);
+                pReco[uiX] = ClipC(static_cast<Short>(pPred[uiX]) + static_cast<Short>(pResi[uiX]));
                 pRecQt[uiX] = pReco[uiX];
                 pRecIPred[uiX] = pReco[uiX];
             }
